Added Z coordinate

+++ DrifterEngine/src/Events/SendFloatingMessageEvent.h
@@ -9,7 +9,7 @@ namespace drft::events
 		std::string message = "INVALID";
 		sf::Color color = sf::Color::Magenta;
 		entt::entity tracksEntity = entt::null;
-		sf::Vector2i position;
+		sf::Vector2f position;
 		sf::Vector2f velocity;
 		bool fades = false;
 		bool isScreenSpace = true;
